move the hash_table impl structs to their own .h file
